Add PartialEq/Eq, Add timeunitspec_eq (#78)
Makes it possible to compare two `Schedule`s for equality.
By default, `Schedule` equality compares the original expression
that was parsed; this test is cheap and provides a holistic
definition of equality. `timeunitspec_eq`, on the other hand,
will consider two `Schedule`s equal if they represent the same
set of `DateTime`s, even if the source expressions were different.